Fix integer handling (mostly)
I still lack support for int64. I hadn't realized before that the C int
was 32bit (on my system at least). I may need to explicitly use int64.
Or maybe this should be more tightly managed by morloc?

I do like the idea of having a safe Int value with a reasonable default.
Problem is, that 32bit is not really reasonable. It isn't high enough to
number the people on earth or the pennies in a big company (and finance
people do use penny values rather than dollars, since fractionals
numbers are evil). So Int64 makes sense. Int64 is big enough that you
are not likely to overflow it unless you are doing number theory. And
you deserve your overflow.

So definitely want to go 64bit. However, that is a lot of memory. It
certainly makes sense to have a 32bit option. And 16, 8, etc. MsgPack
does not go to 128, nor does it really need to. The only time you'd need
128 bit integers is if you are doing crypto or math. And then you'll
also need 256 adn 512 etc. At that point, you might as well just jump to
binary or big ints.