I have found that I am a goose
Somewhere I went off in the wrong direction with my parsing. The
mpack_read function loses context and gets terribly lost. I that I
should be using the mpack_parse and mpack_unparse functions with their
callback support. There is an example of JSON parse/unparse on the main
repo that shows how this works.