The python interop is working now
And I have a very nice way to make schemas and encode data. This makes
testing very easy through python. And it is good that I am testing,
because the parser is borken badly. Numbers of 0xffff are falling to 0.
The tokenizer seems to be getting a byte or two offset after that. This
causes all following tokens to be foobared. And dictionaries are
segfaulting under seemingly random conditions.