Replace C schema structure with string schema
The string schema has the following grammar:

```
 schema = 'a' schema                -- array
        | 't' n [schema]_n          -- tuple
        | 'm' n [n string schema]_n -- MAP
        | 'z'                       -- NIL
        | 'b'                       -- BOOL
        | 'i' w                     -- SINT
        | 'u' w                     -- UINT
        | 'f' w                     -- FLOAT
        | 's'                       -- STRING
        | 'r'                       -- BINARY
```

where n is the size of the container (tuple or map) and w is the width
of the number in bytes. Numbers are encoded in base 64 with the
following mapping:

 characters 0-9    --> integers 0-9
 characters a-z    --> integers 10-35
 characters A-Z    --> integers 36-61
 '+'               --> 62
 '/'               --> 63

In the future, I will extend this with one additional character, '=',
that can be used to create numbers larger than 63. For now, this not
really critical since I am quite happy to disallow map fields, tuple
sizes, and record sizes of over 63. Almost, I can image records that
will need to be large, for example large config files.

For the integer mappings, I currently only support i32 and f64.
Extending to the full set, i8, i16, i32, i64, u8, u16, u32, and u64 is
certainly a priority, but I can do this after the paper is resubmitted.