rm joy, my C++ test weren't and my py test aren't
So my C++ tests was just checking for success of packing and unpacking.
It wasn't checking the ultimate values. This was a decision I made very
early since I was then focusing on just getting the damn thing to
compile and not segfault. But then I forgot to circle back and test the
values. And, well, C++ is actually failing hard. The arrays are all
wrong, among other things.

My python tests were mostly correct in circling to voidstar and back.
But this does not mean that they were going to the right place. Ugh, I
am not in the good place.