Long lists of integers fail in python
When 2000 integers are processed in python (e..g, list(range(2000))),
part way through there seems to be a skip where one integer is not
processed and all following ones are off by one. Running the same data
in C++ does not do this, so it seems to be a python-only problem.