Fixed bug in token length
The token length was not being set for non-float primitives. When I
attempted to resize the buffer using the token length for the required
additional space, an unset address was being resolved. For unoptimized
code this was random memory, but, for some reason, with the -O option it
was always zero. On zero it failed, while most random memory locations
where positive.