Allow maps - all R tests pass
Currently maps are treated as named lists in R. This is not very
efficient, since lookup times for keys in a named list run in linear
time. That is, lists are not implemented as hashes or trees. They are
simply vectors with a character vector of names associated.

R environments are more efficient data structures, though this
improvement is minimal for short lists, and since named lists are
usually representing records, usually of fewer than tens of fields, the
cost of linear lookup is minimal and may even be faster than lookup in
more advanced structures.

All R tests now pass.

The TODO list is now:
 [ ] write more extensive R tests
 [ ] integrate all this back into morloc
 [ ] rewrite mlcmpack to use shared memory
 [ ] benchmark and pray to the silicon gods that the shm solution is
     faster and doesn't blow up the memory
 [ ] update all libraries to use the new numeric system
 [ ] update the case studies
 [ ] update the paper with the new architecture
 [ ] publish the paper
 [ ] update all documentation (as the paper makes it way through the
     last editorial process)