Add more tests, fix long string bug
I had been thinking that the largest integer R can express is 2^54-1. This
is the case since large integers are stored as doubles. Doubles in
IEEE754 are stored as 64bit values with 1 sign bit, 11 exponent bits,
and 53 precision bits. But there seem to be lower limits on integers set
for at least some functions. For sample.int, a custom error is raised
for integers above the exact integer 4,500,000,000,000,000.

As far as I can tell, the actual R integer type is a 32bit int that is
automatically converted to a double when incremented above/below
(+/-)2^31. This is only slightly odd, since I would expect the value
-2^31 to be within the i32 range, which should run from -2^31 to 2^31-1.